refactor(activerecord): relocate find/findBy/findByBang to core.ts (M2c)
